requirements.txt
fastapi
uvicorn
pandas
scikit-learn
joblib
sentence-transformers
python-jose[cryptography]
requests
python-multipart
python-dotenv
groq
